Gather all controllers specs in one file as required
