+++ Gemfile
@@ -39,3 +39,5 @@ group :test do
 end
 
 gem 'rubocop', '>= 1.0', '< 2.0'
+
+gem 'rails-controller-testing'
